Implemented a more rigorous system for determining the file path of instances.  Instances that are not in the same directory as the parent cell will be written out to the parent cell definition as "use <cellname> <id> <path>" instead of "use <cellname> <id>". This is intended to prevent confusion between multiple versions of a cell definition that are all in the search path.
To go along with this, the "cellname" command has been extended
to include the command option "cellname filepath" that can be
used either to return the name of the file path for the cell,
or to set or change the file path.  Removing the file path causes
magic to revert to the former behavior of looking in the whole
search path for the cell.  The file path can be removed using
"cellname filepath <cellname> default".